// src/models/Enums/ESkillAction.ts
export enum ESkillAction{
    NONE='None',
    HEROATTACK = 'Hero Attack',
    BURSTDAMAGE = 'Burst Damage',
    TICKINGDAMAGE = 'Ticking Damage',

    HEAL = 'Heal',
    MANAHEAL = 'Mana Heal',
    TICKINGHEAL = 'Ticking Heal',

    ARMORBUFF = 'Armor Buff',
    EVASIONBUFF = 'Evasion Buff',
    HERODAMAGEBUFF = 'Hero Damage Debuff',
    MAGICRESISTANCEBUFF = 'Magic Resistance Buff',

    ARMORDEBUFF = 'Armor Debuff',
    EVASIONDEBUFF = 'Evasion DeBuff',
    ENEMYDAMAGEDEBUFF = 'Enemy Damage Debuff',
    MAGICRESISTANCEDEBUFF = 'Magic Resistance Debuff',

    STUN = 'Stun',
    SLOWDOWN = ' Slow Down'

}